requirements.txt
torch
torchvision
torchaudio
pandas
scikit-learn
matplotlib
seaborn
tqdm
Pillow